Smaller 8Kx8 EEPROM is working
I managed to get the code to generate a ROM file which uses a 4-bit board
state plus the user's moves to look up the net board move and the next
board state. I've redone the schematic, PCB and I've started on the
documentation. I need to redraw the design diagram.